refactor: consolidate distance utilities
